pyproject.toml
[project]
name = "nabu-agent"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faster-whisper>=1.2.0",
    "langchain-community>=0.3.26",
    "langchain-mcp-adapters>=0.1.10",
    "langchain-openai>=0.3.33",
    "langgraph>=0.4.8",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "spotipy>=2.25.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[project.scripts]
nabu-agent = "nabu_agent:main.app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
